Rendben működik a label texting, és a logolás is.

+++ src/main/java/classes/Ai/Evaluation/Evaluator.java
@@ -13,7 +13,7 @@ public class Evaluator {
 
     //region With FieldValues
 
-    public static double evaluate(int possibilityNumAfterMove, int enemyKingPossibilityNumAfterMove, boolean forWhite){
+    public static double evaluate(int possibilityNumAfterMove, double enemyKingPossibilityNumAfterMove, boolean forWhite){
         double eval = evaluateWithFieldValues();
         eval += (forWhite ? +1 : -1) * 0.1 * possibilityNumAfterMove;
         if (currentGameState == SECOND_STATE)
